Unify treatment of sensors and replays
