+++ jumpstreet/configs/sensors/MOT15-replay.yml
@@ -1,14 +1,30 @@
 ---
 type: "camera"
-model: "replay"
+sensor_class: "ReplayCamera"
 dataset: "ADL-Rundle-6"
 data_path: "data/MOT15/train/ADL-Rundle-6/img1"
 serial: Null
 width: 1920
 height: 1080
 fps: 30
+jpg_compression_pct: 80
 interpolate: Null
 io_threads: 2
+calibration:
+  channel_order: "bgr"
+  intrinsics:
+    fx: 1448  # unsure...
+    fy: 1448  # unsure...
+    u: 960  # width / 2
+    v: 512  # height / 2
+    g: 0
+  extrinsics:
+    dx: 0
+    dy: 0
+    dz: 0
+    qx: 0
+    qy: 0
+    qz: 0
 backend:
   transport: "ipc"
   host: "localhost"
